gen-ai/how-to-guides/trace_your_pipeline/pyproject.toml
[project]
name = "trace-your-pipeline"
version = "0.0.0"
description = "Trace your pipeline example"
requires-python = ">=3.11,<3.13"
readme = "README.md"
dependencies = [
    "gllm-core==0.4.24",
    "gllm-inference==0.6.29",
    "gllm-pipeline==0.5.8",
    "opentelemetry-sdk==1.41.1",
]

[[tool.uv.index]]
name = "gen-ai-internal"
url = "https://glsdk.gdplabs.id/gen-ai-internal/simple/"

[tool.uv.sources]
gllm-core = { index = "gen-ai-internal" }
gllm-inference = { index = "gen-ai-internal" }
gllm-pipeline = { index = "gen-ai-internal" }
